fix(rxSortableColumn): user proper ID to pass function to scope
BREAKING CHANGE: The sortMethod of rxSortableColumn was not properly
passing the method to the `scope` with a `&`, but was using a `=`, which
should only be used for two-way bindings.

To migrate the code, follow the example below:

BEFORE:

<rx-sortable-column
    sort-method="sortCol"
    sort-property="name"
    predicate="sort.predicate"
    reverse="sort.reverse">

AFTER:

<rx-sortable-column
    sort-method="sortCol(property)"
    sort-property="name"
    predicate="sort.predicate"
    reverse="sort.reverse">

The method you pass will *always* take `property` as the argument
when assigning it to `sort-method`.

tl;dr: Change every `sort-method="foo"` assignment to
`sort-method="foo(property)"`