fix(rxForm): Ensure strings come out of `getContent`
Closes #294

The result of `getContent` is passed to `ng-bind-html`. This directive
*requires* "safe" HTML. If the value of `row[expr]` is an integer,
then `ng-bind-html` causes an exception to be thrown, as that's not
valid safe HTML. Converting it to a string solves this.

In Cloud and other projects, we had people doing things like:
`key: '{{ram}}'`, instead of `key: 'ram'`, because otherwise
nothing was being shown on screen. No one understood why, so we all
just did `'{{ram}}'`, as that "fixed" it. It fixed it because it caused
everything to get passed to `$interpolate`, which creates a string...