test(rxNotify): Page objects, midway tests.
Closes #330

There is a test in here for making sure that a notification dismisses
itself successfully after X seconds, which worked on my machine. If my
machine is able to manipulate the DOM fast enough, it will look like it
works fine, but on slower machines (namely TravisCI), there may be issues.