feat(statusTags): Implement page/breadcrumb tags
Some of our products have been putting ALPHA and BETA tags
into page titles and breadcrumbs by making use of ng-bind-html
in the appropriate name/title fields. This PR formalizes that
as part of the framework.

We've added optional labels to `<rx-page>` via the `status` attribute,
and to breadcrumbs via the `status` attribute.

The framework currently ships with `"alpha"` and `"beta"` tags. If there
is demand for more general ones in the future, we can add those.

This also adds support for per-product custom tags. Products can use
`rxStatusTagsProvider.addStatus` in their `.config()` methods, and then
freely use that status tag in their product.

Closes #638