fix(rxSortableColumn): Put sort arrow in <button>
This commit puts the sort icon into the <button> that rxSortableColumn
creates, allowing us to click on the sort arrow itself.

This also updates rxStatusColumn to allow for sorting of the status
column. It adds documentation to the rxStatusColumn demo page to show
how to do this.

Finally, we also fixed a problem where tooltips would not show for
ACTIVE and PENDING status column cells. It was slightly related to
the sorting issue, so it has been bundled in to this commit.

Closes #732
Closes #733