test(rxCharacterCount): Include a "main" instance
This is there to make sure that if you only have one rxCharacterCount
element on a page, you can quickly get to it by using a typical selector
from the template. In this case, it's the html attribute found using the
CSS selector `$(['rx-character-class'])`.

Also, I missed some documentation linking up stuff in the template
generator, so I fixed it in both places. My bad.