test(rxCharacterCount): Exercises
These exercises are a significant amount of work, and for that I
apologize. But testers need them to make sure that a component is
working at a low-level without writing a bunch of tests. So, the time
you spend writing exercises saves time down the line by huge amounts for
new apps.

Basically, the strategy is to define some defaults (which were pretty
easy with your documentation in the demo page), and then copy/paste the
existing midway tests into the exercises file. From there, you start
pruning the tests that can't be run/should be run if certain options
exist/don't exist/meet some sort of predefined value. These need to be
documented as well since they end up in the page object's API docs.

I also forgot some things in the template. Imagine that.