fix(rxCharacterCount): Clean up the generated elements
Closes #847

Properly clean up the character count and other generated elements,
created when rx-character-count is used on an element.

This also moves the character count into the `.counted-input-wrapper`
div. Previously it sat outside of it.

BREAKING CHANGE
We've removed the `.character-count-wrapper` class. Previously we
set widths on this class. It ended up being redundant, widths
should now be set w.r.t. `.counted-input-wrapper`