fix(rxForm): Remove variables that reference DOM
In Selenium, storing an element into a variable only stores a reference
to the variable, meaning that if Angular were to destroy the node the
element was in (perhaps using `ng-if`) and repaint it, the variable
would still have the old reference.

http://docs.seleniumhq.org/exceptions/stale_element_reference.jsp

To avoid this, never assign an element to a variable. Instead, only use
functions that, when invoked, explicitly visit the most recent DOM to
fetch an element, every single time.