fix(visualRegression): Correct branch names
The idea is to have visual regression branch names match the SHA of the
commit that Travis just got done building. I thought it made sense to
grab this commit SHA directly from git itself, but it appears that
Travis only clones a portion of the repository or something. It reports
back different SHAs for the commits than what's in github.

From now on, grab the SHA from the range of commits that Travis sees
from github, which is located in a special Travis environment variable.