test(rxMetadata): Account for missing transformFns
Should someone use a `main` instance of rxMetadata, there will
be an `undefined` return value for `transformFns`. This throws an
exception later when a term is requested, since it need to check if
the term exists in the transform object. Searching for a key in
something that is `undefined` throws an error.